api.orders/requirements.txt
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.0
Flask==2.2.3
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.0.11
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
peewee==3.15.4
pluggy==1.0.0
pytest==7.2.1
tomli==2.0.1
Werkzeug==2.2.3
zipp==3.13.0
python-dotenv
flask-cors
